Adding support for providing the token via header.
This is useful for XHR requests from the JavaScript to the server, where you typically send JSON instead of form encoded data.

Ruby on Rails has a built-in forgery protection similar to ring-anti-forgery, and also allows for setting the token via a header (x-csrf-token), so this is a proven method.